+++ tests/test_plot/test_drill.py
@@ -23,16 +23,16 @@ DRILL_DIR = 'Drill'
 positions = {'R1': (105, 35, 'top'), 'R2': (110, 35, 'bottom'), 'R3': (110, 45, 'top')}
 
 
-def do_3Rs(conf, dir, modern):
+def do_3Rs(conf, dir, modern, single=False):
     ctx = context.TestContext(dir, '3Rs', conf, DRILL_DIR)
     ctx.run()
     # Check all outputs are there
     pth_drl = ctx.get_pth_drl_filename()
     npth_drl = ctx.get_npth_drl_filename()
-    pth_gbr_drl = ctx.get_pth_gbr_drl_filename()
-    npth_gbr_drl = ctx.get_npth_gbr_drl_filename()
     pth_pdf_drl = ctx.get_pth_pdf_drl_filename()
     npth_pdf_drl = ctx.get_npth_pdf_drl_filename()
+    pth_gbr_drl = ctx.get_pth_gbr_drl_filename()
+    npth_gbr_drl = ctx.get_npth_gbr_drl_filename()
     report = 'report.rpt'
 
     if modern:
@@ -43,6 +43,16 @@ def do_3Rs(conf, dir, modern):
         pth_pdf_drl = pth_pdf_drl.replace('-drl', '_drill')
         npth_pdf_drl = npth_pdf_drl.replace('-drl', '_drill')
         report = '3Rs-drill_report.txt'
+        if single:
+            pth_drl = pth_drl.replace('PTH_', '')
+            npth_drl = npth_drl.replace('NPTH_', '')
+            pth_pdf_drl = pth_pdf_drl.replace('PTH_', '')
+            npth_pdf_drl = npth_pdf_drl.replace('NPTH_', '')
+    elif single:
+        pth_drl = pth_drl.replace('-PTH', '')
+        npth_drl = npth_drl.replace('-NPTH', '')
+        pth_pdf_drl = pth_pdf_drl.replace('-PTH', '')
+        npth_pdf_drl = npth_pdf_drl.replace('-NPTH', '')
 
     ctx.expect_out_file(os.path.join(DRILL_DIR, report))
     ctx.expect_out_file(pth_drl)
@@ -56,7 +66,7 @@ def do_3Rs(conf, dir, modern):
     ctx.expect_gerber_flash_at(pth_gbr_drl, 6, (110, -45))
     ctx.expect_gerber_has_apertures(pth_gbr_drl, ['C,1.000000'])
     # We have a mounting hole at (120, 29) is 2.1 mm in diameter
-    ctx.search_in_file(npth_drl, ['X120.0Y-29.0', 'T1C2.100'])
+    ctx.search_in_file(npth_drl, ['X120.0Y-29.0', 'T.C2.100'])
     ctx.expect_gerber_flash_at(npth_gbr_drl, 6, (120, -29))
     ctx.expect_gerber_has_apertures(npth_gbr_drl, ['C,2.100000'])
     ctx.clean_up()
@@ -66,5 +76,13 @@ def test_drill_3Rs():
     do_3Rs('drill', 'test_drill_3Rs', True)
 
 
+def test_drill_single_3Rs():
+    do_3Rs('drill_single', 'test_drill_single_3Rs', True, True)
+
+
 def test_drill_legacy_3Rs():
     do_3Rs('drill_legacy', 'test_drill_legacy_3Rs', False)
+
+
+def test_drill_legacy_s_3Rs():
+    do_3Rs('drill_legacy_s', 'test_drill_legacy_s_3Rs', False, True)