+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Problems when using the `pdf_pcb_print` output and variants to remove
   a component with ridiculous pads that only has solder paste (no
   copper, nor even solder mask aperture).
+- Excellon drill output when using unified output and not using default
+  KiCad names.
 
 
 ## [0.7.0] - 2020-09-11